backend/requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
flask
flask-cors
fpdf2
gunicorn
